Enables big-endian architecture support (#86)
* Adds big-endian support

Adds endianness detection and byte-swap macros to
ensure correct data interpretation on big-endian
architectures.

Implements byte-swapping for 16, 32, and 64-bit
values and integrates them into the read/store
functions to handle potential endianness differences.

* Adds endian-agnostic partial read function

Introduces `zxc_le_partial` to load partial data in
little-endian order, ensuring consistent behavior across
architectures. This removes endian-dependent memcpy.

Updates bit reader initialization and refill logic to use
`zxc_le_partial`, improving portability and correctness,
particularly on big-endian systems.

* Increases size of compress bound return value

Updates the `zxc_compress_bound` function to return a `uint64_t` instead of a `size_t` to accommodate larger input sizes and prevent potential overflow issues on platforms with smaller `size_t` implementations. This change also affects the `zxc_get_decompressed_size` function, updating its return type to `uint64_t` as well.

* Fix: casts lengths to correct types

Casts the length of data buffers to the correct type to ensure compatibility and prevent potential overflow issues, particularly when dealing with larger data sizes.

* Adds multi-architecture build workflow

Introduces a new GitHub Actions workflow for building and testing the library across multiple architectures.

This workflow categorizes architectures into three tiers based on their support level:
- Tier 1: Main architectures, built on releases, manual triggers, and pull requests.
- Tier 2: Extended Linux architectures, built on releases and manual triggers.
- Tier 3: Experimental Linux architectures, built on manual triggers only.

The workflow uses QEMU for emulation and cross-compilers for building.

* Adds big-endian support

Implements support for big-endian architectures by adjusting memory operations.

Adds big-endian support for decoding

Adds a conditional compilation block to handle big-endian architectures when storing decoded values.

This ensures correct byte order during decompression on big-endian systems.

* Enables position independent code and fixes LE test

Enables position independent code for shared libraries.

Addresses a potential test failure on big-endian architectures by ensuring correct byte order in a test assertion. This change aligns the expected accumulated value with the little-endian interpretation of the input buffer.

* Improves repetitive pattern encoding test

Updates the repetitive pattern encoding test to use a 5-byte
period to avoid NUM data classification and exercise 8-bit offset
encoding. This ensures proper testing of small offset encoding
functionality, especially important for big-endian systems.

Adds a warning for when endianness is not detected during compilation.

* Refactors build multiarch configurations

Restructures build configurations in the workflow file to improve readability and maintainability.

Removes conditional execution for Tier 2 and Tier 3 builds, indicating they are now consistently built.

Adds Linux loong64 architecture to Tier 3, enabling testing on this platform.

* Renames and updates multiarch workflow

Renames the build workflow to better reflect its purpose.

* Fixes big-endian handling and checksum

Corrects byte increment in big-endian code to use `sizeof`
instead of hardcoded values.

Also corrects an incorrect checksum index in the test suite.

Updates rust wrapper to cast compress bound to usize to be able to create a vector of that size.

* Fix typo

+++ include/zxc_buffer.h
@@ -9,6 +9,7 @@
 #define ZXC_BUFFER_H
 
 #include <stddef.h>
+#include <stdint.h>
 
 #include "zxc_export.h"
 
@@ -29,7 +30,7 @@
  *
  * @return           Maximum required buffer size in bytes.
  */
-ZXC_EXPORT size_t zxc_compress_bound(const size_t input_size);
+ZXC_EXPORT uint64_t zxc_compress_bound(const size_t input_size);
 
 /**
  * @brief Compresses a data buffer using the ZXC algorithm.
@@ -85,6 +86,6 @@ ZXC_EXPORT size_t zxc_decompress(const void* src, const size_t src_size, void* d
  * @return The original uncompressed size in bytes, or 0 if the buffer is invalid
  *         or too small to contain a valid ZXC archive.
  */
-ZXC_EXPORT size_t zxc_get_decompressed_size(const void* src, const size_t src_size);
+ZXC_EXPORT uint64_t zxc_get_decompressed_size(const void* src, const size_t src_size);
 
 #endif  // ZXC_BUFFER_H
